feat(backend): updated the notifaction message handler
Updated the notifaction message handler to cope w/ proper CAP Alert
formatting.  Moved the cap_area_to_geojson to the core module, as it is
used by both notifications and alerts.
