+++ server/safers/core/utils/__init__.py
@@ -1,4 +1,5 @@
 from .utils_settings import DynamicSetting
-from .utils_urls import remove_urlpatterns, DateTimeConverter
 from .utils_enums import CaseInsensitiveTextChoices
+from .utils_geometry import cap_area_to_geojson
+from .utils_urls import remove_urlpatterns, DateTimeConverter
 from .utils_validators import validate_reserved_words, validate_schema
